IDVA6-531 Add company tests separated to their own file, new tests for invitations banner added and mocks updated
