+++ test/mocks/associations.mock.ts
@@ -21,6 +21,26 @@ export const userAssociations: Associations = {
     ]
 } as Associations;
 
+export const userAssociationsWithNumberOfInvitations: Associations = {
+    items: [
+        {
+            id: "1234567890",
+            userId: "qwertyiop",
+            userEmail: "demo@ch.gov.uk",
+            companyNumber: "NI038379",
+            companyName: "THE POLISH BREWERY"
+        },
+        {
+            id: "2345678901",
+            userId: "qwertyiop",
+            userEmail: "demo@ch.gov.uk",
+            companyNumber: "01777777",
+            companyName: "BRITISH AIRWAYS PLC"
+        }
+    ],
+    totalResults: 3
+} as Associations;
+
 export const companyAssociations: Associations = {
     items: [
         {
