+++ test/src/routers/controllers/yourCompaniesController.addCompany.test.ts
@@ -1,0 +1,287 @@
+// the order of the mock imports matter
+import mocks from "../../../mocks/all.middleware.mock";
+import { validActiveCompanyProfile, validDisolvedCompanyProfile } from "../../../mocks/companyProfile.mock";
+import { Resource } from "@companieshouse/api-sdk-node";
+import { CompanyProfile } from "@companieshouse/api-sdk-node/dist/services/company-profile/types";
+import app from "../../../../src/app";
+import supertest from "supertest";
+import { StatusCodes } from "http-status-codes";
+import * as commpanyProfileService from "../../../../src/services/companyProfileService";
+import * as associationService from "../../../../src/services/userCompanyAssociationService";
+import errorManifest from "../../../../src/lib/utils/error_manifests/errorManifest";
+import { COMPNANY_ASSOCIATED_WITH_USER, COMPNANY_NOT_ASSOCIATED_WITH_USER } from "../../../../src/constants";
+
+const router = supertest(app);
+
+jest.mock("../../../../src/lib/Logger");
+jest.mock("../../../../src/lib/utils/sessionUtils", () => {
+    const originalModule = jest.requireActual("../../../../src/lib/utils/sessionUtils");
+
+    return {
+        __esModule: true,
+        ...originalModule,
+        getLoggedInUserEmail: jest.fn(() => "test@test.com")
+    };
+});
+
+it("should check session and auth before returning the add company page", async () => {
+    await router.get("/your-companies/add-company");
+    expect(mocks.mockSessionMiddleware).toHaveBeenCalled();
+    expect(mocks.mockAuthenticationMiddleware).toHaveBeenCalled();
+});
+
+describe("GET /your-companies/add-company", () => {
+    const en = require("../../../../src/locales/en/translation/add-company.json");
+    const cy = require("../../../../src/locales/cy/translation/add-company.json");
+    const enCommon = require("../../../../src/locales/en/translation/common.json");
+    const cyCommon = require("../../../../src/locales/cy/translation/common.json");
+
+    beforeEach(() => {
+        jest.clearAllMocks();
+    });
+
+    it("should return status 200", async () => {
+        await router.get("/your-companies/add-company").expect(200);
+    });
+
+    it("should return expected English content if language version set to English", async () => {
+        const response = await router.get("/your-companies/add-company?lang=en");
+        expect(response.text).toContain(enCommon.back_link);
+        expect(response.text).toContain(en.what_is_the_company_number);
+        expect(response.text).toContain(en.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(en.you_can_find_this_by_searching);
+        expect(response.text).toContain(en.how_do_i_find_the_company_number);
+        expect(response.text).toContain(enCommon.continue);
+    });
+
+    it("should return expected Welsh content if language version set to Welsh", async () => {
+        const response = await router.get("/your-companies/add-company?lang=cy");
+        expect(response.text).toContain(cyCommon.back_link);
+        expect(response.text).toContain(cy.what_is_the_company_number);
+        expect(response.text).toContain(cy.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(cy.you_can_find_this_by_searching);
+        expect(response.text).toContain(cy.how_do_i_find_the_company_number);
+        expect(response.text).toContain(cyCommon.continue);
+    });
+});
+
+describe("POST /your-companies/add-company", () => {
+    const en = require("../../../../src/locales/en/translation/add-company.json");
+    const cy = require("../../../../src/locales/cy/translation/add-company.json");
+    const enCommon = require("../../../../src/locales/en/translation/common.json");
+    const cyCommon = require("../../../../src/locales/cy/translation/common.json");
+
+    beforeEach(() => {
+        jest.clearAllMocks();
+    });
+
+    it("should return status 302 if company number is correct and the company is active and not already associated with the user", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockReturnValue(validActiveCompanyProfile);
+        const associationSpy: jest.SpyInstance = jest.spyOn(associationService, "isCompanyAssociatedWithUser");
+        associationSpy.mockReturnValue(COMPNANY_NOT_ASSOCIATED_WITH_USER);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=en").send({ companyNumber: "12345678" });
+        // Then
+        expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    });
+
+    it("should return expected English error message if company number is correct but company status is not active and language version set to English", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockReturnValue(validDisolvedCompanyProfile);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=en").send({ companyNumber: "23456789" });
+        // Then
+        expect(response.text).toContain(enCommon.back_link);
+        expect(response.text).toContain(enCommon.there_is_a_problem);
+        expect(response.text).toContain(en.what_is_the_company_number);
+        expect(response.text).toContain(en.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(en.you_can_find_this_by_searching);
+        expect(response.text).toContain(en.how_do_i_find_the_company_number);
+        expect(response.text).toContain(enCommon.continue);
+        expect(response.text).toContain(en.enter_a_company_number_for_a_company_that_is_active);
+    });
+
+    it("should return expected Welsh error message if company number is correct but company status is not active and language version set to Welsh", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockReturnValue(validDisolvedCompanyProfile);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=cy").send({ companyNumber: "23456789" });
+        // Then
+        expect(response.text).toContain(cyCommon.back_link);
+        expect(response.text).toContain(cyCommon.there_is_a_problem);
+        expect(response.text).toContain(cy.what_is_the_company_number);
+        expect(response.text).toContain(cy.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(cy.you_can_find_this_by_searching);
+        expect(response.text).toContain(cy.how_do_i_find_the_company_number);
+        expect(response.text).toContain(cyCommon.continue);
+        expect(response.text).toContain(cy.enter_a_company_number_for_a_company_that_is_active);
+    });
+
+    it("should return expected English error message if company number is correct but company has already been added to user account and language version set to English", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockReturnValue(validActiveCompanyProfile);
+        const associationSpy: jest.SpyInstance = jest.spyOn(associationService, "isCompanyAssociatedWithUser");
+        associationSpy.mockReturnValue(COMPNANY_ASSOCIATED_WITH_USER);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=en").send({ companyNumber: "12345678" });
+        // Then
+        expect(response.text).toContain(enCommon.back_link);
+        expect(response.text).toContain(enCommon.there_is_a_problem);
+        expect(response.text).toContain(en.what_is_the_company_number);
+        expect(response.text).toContain(en.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(en.you_can_find_this_by_searching);
+        expect(response.text).toContain(en.how_do_i_find_the_company_number);
+        expect(response.text).toContain(enCommon.continue);
+        expect(response.text).toContain(en.this_company_has_already_been_added_to_your_account);
+    });
+
+    it("should return expected Welsh error message if company number is correct but company has already been added to user account and language version set to Welsh", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockReturnValue(validActiveCompanyProfile);
+        const associationSpy: jest.SpyInstance = jest.spyOn(associationService, "isCompanyAssociatedWithUser");
+        associationSpy.mockReturnValue(COMPNANY_ASSOCIATED_WITH_USER);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=cy").send({ companyNumber: "12345678" });
+        // Then
+        expect(response.text).toContain(cyCommon.back_link);
+        expect(response.text).toContain(cyCommon.there_is_a_problem);
+        expect(response.text).toContain(cy.what_is_the_company_number);
+        expect(response.text).toContain(cy.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(cy.you_can_find_this_by_searching);
+        expect(response.text).toContain(cy.how_do_i_find_the_company_number);
+        expect(response.text).toContain(cyCommon.continue);
+        expect(response.text).toContain(cy.this_company_has_already_been_added_to_your_account);
+    });
+
+    it("should return expected English error message if company number is incorrect and language version set to English", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockRejectedValue({
+            httpStatusCode: StatusCodes.BAD_REQUEST
+        } as Resource<CompanyProfile>);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=en").send({ companyNumber: "%£$£$£5343" });
+        // Then
+        expect(response.text).toContain(enCommon.back_link);
+        expect(response.text).toContain(enCommon.there_is_a_problem);
+        expect(response.text).toContain(en.what_is_the_company_number);
+        expect(response.text).toContain(en.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(en.you_can_find_this_by_searching);
+        expect(response.text).toContain(en.how_do_i_find_the_company_number);
+        expect(response.text).toContain(enCommon.continue);
+        expect(response.text).toContain(en.enter_a_company_number_that_is_8_characters_long);
+    });
+
+    it("should return expected Welsh error message if company number is incorrect and language version set to Welsh", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockRejectedValue({
+            httpStatusCode: StatusCodes.BAD_REQUEST
+        } as Resource<CompanyProfile>);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=cy").send({ companyNumber: "%£$£$£5343" });
+        // Then
+        expect(response.text).toContain(cyCommon.back_link);
+        expect(response.text).toContain(cyCommon.there_is_a_problem);
+        expect(response.text).toContain(cy.what_is_the_company_number);
+        expect(response.text).toContain(cy.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(cy.you_can_find_this_by_searching);
+        expect(response.text).toContain(cy.how_do_i_find_the_company_number);
+        expect(response.text).toContain(cyCommon.continue);
+        expect(response.text).toContain(cy.enter_a_company_number_that_is_8_characters_long);
+    });
+
+    it("should return expected English error message if there is no company with provided number and language version set to English", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockRejectedValue({
+            httpStatusCode: StatusCodes.NOT_FOUND
+        } as Resource<CompanyProfile>);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=en").send({ companyNumber: "11111111" });
+        // Then
+        expect(response.text).toContain(enCommon.back_link);
+        expect(response.text).toContain(enCommon.there_is_a_problem);
+        expect(response.text).toContain(en.what_is_the_company_number);
+        expect(response.text).toContain(en.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(en.you_can_find_this_by_searching);
+        expect(response.text).toContain(en.how_do_i_find_the_company_number);
+        expect(response.text).toContain(enCommon.continue);
+        expect(response.text).toContain(en.enter_a_company_number_that_is_8_characters_long);
+    });
+
+    it("should return expected Welsh error message if there is no company with provided number and language version set to Welsh", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockRejectedValue({
+            httpStatusCode: StatusCodes.NOT_FOUND
+        } as Resource<CompanyProfile>);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=cy").send({ companyNumber: "11111111" });
+        // Then
+        expect(response.text).toContain(cyCommon.back_link);
+        expect(response.text).toContain(cyCommon.there_is_a_problem);
+        expect(response.text).toContain(cy.what_is_the_company_number);
+        expect(response.text).toContain(cy.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(cy.you_can_find_this_by_searching);
+        expect(response.text).toContain(cy.how_do_i_find_the_company_number);
+        expect(response.text).toContain(cyCommon.continue);
+        expect(response.text).toContain(cy.enter_a_company_number_that_is_8_characters_long);
+    });
+
+    it("should return expected English error message if there is no company number provided and language version set to English", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockRejectedValue({
+            httpStatusCode: StatusCodes.NOT_FOUND
+        } as Resource<CompanyProfile>);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=en").send({ companyNumber: "" });
+        // Then
+        expect(response.text).toContain(enCommon.back_link);
+        expect(response.text).toContain(enCommon.there_is_a_problem);
+        expect(response.text).toContain(en.what_is_the_company_number);
+        expect(response.text).toContain(en.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(en.you_can_find_this_by_searching);
+        expect(response.text).toContain(en.how_do_i_find_the_company_number);
+        expect(response.text).toContain(enCommon.continue);
+        expect(response.text).toContain(en.enter_a_company_number);
+    });
+
+    it("should return expected Welsh error message if there is no company number provided and language version set to Welsh", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockRejectedValue({
+            httpStatusCode: StatusCodes.NOT_FOUND
+        } as Resource<CompanyProfile>);
+        // When
+        const response = await router.post("/your-companies/add-company?lang=cy").send({ companyNumber: "" });
+        // Then
+        expect(response.text).toContain(cyCommon.back_link);
+        expect(response.text).toContain(cyCommon.there_is_a_problem);
+        expect(response.text).toContain(cy.what_is_the_company_number);
+        expect(response.text).toContain(cy.a_company_number_is_8_characters_long);
+        expect(response.text).toContain(cy.you_can_find_this_by_searching);
+        expect(response.text).toContain(cy.how_do_i_find_the_company_number);
+        expect(response.text).toContain(cyCommon.continue);
+        expect(response.text).toContain(cy.enter_a_company_number);
+    });
+
+    it("should return expected generic error message if any other error happens", async () => {
+        // Given
+        const companyProfileSpy: jest.SpyInstance = jest.spyOn(commpanyProfileService, "getCompanyProfile");
+        companyProfileSpy.mockRejectedValue({
+            httpStatusCode: StatusCodes.BAD_GATEWAY
+        } as Resource<CompanyProfile>);
+        // When
+        const response = await router.post("/your-companies/add-company").send({ companyNumber: "12345678" });
+        // Then
+        expect(response.text).toContain(errorManifest.generic.serverError.summary);
+    });
+});
